% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDTreeAlgorithm.R
\name{KDimensionalTree}
\alias{KDimensionalTree}
\title{Create a K-dimensional tree object.}
\usage{
KDimensionalTree(points, distanceFunction = NULL)
}
\arguments{
\item{points}{A numerical matrix, each row corresponding to a point.}

\item{distanceFunction}{Distance function for to use.
If NULL then Euclidean distance is used.}
}
\description{
Contructs a K-dimensional tree S3 object based on given points.
}
