% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDTreeAlgorithm.R
\name{KNearest}
\alias{KNearest}
\title{Nearest k-points.}
\usage{
KNearest(tree, point, k = 1, format = NULL)
}
\arguments{
\item{tree}{K-dimensional tree S3 object.}

\item{k}{Number of nearest neighbors to find.}

\item{format}{Format of the result.
One NULL, "data.frame", "list", or "points".}

\item{Query}{point.}
}
\description{
For a given point finds the nearest point(s) to it.
}
