% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDTreeAlgorithm.R
\name{NearestWithinBall}
\alias{NearestWithinBall}
\title{Nearest within ball.}
\usage{
NearestWithinBall(tree, point, radius, valuesQ = TRUE)
}
\arguments{
\item{tree}{K-dimensional tree S3 object.}

\item{radius}{Ball radius.}

\item{Query}{point.}

\item{format}{Format of the result.
One NULL, "data.frame", "list", or "points".}
}
\description{
For a given point finds the nearest point(s) to it within a given ball radius.
}
