% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDTreeAlgorithm.R
\name{KNearest}
\alias{KNearest}
\title{Nearest k-points.}
\usage{
KNearest(tree, point, k = 1, valuesQ = TRUE)
}
\arguments{
\item{tree}{K-dimensional tree S3 object.}

\item{k}{Number of nearest neighbors tp find.}

\item{valuesQ}{Should values be returned or not?}

\item{Query}{point.}
}
\description{
For a given point finds the nearest point(s) to it.
}
