% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDTreeAlgorithm.R
\name{NearestWithinBall}
\alias{NearestWithinBall}
\title{Nearest within ball.}
\usage{
NearestWithinBall(tree, point, r, valuesQ = TRUE)
}
\arguments{
\item{tree}{K-dimensional tree S3 object.}

\item{r}{Ball radius.}

\item{valuesQ}{Should values be returned or not?}

\item{Query}{point.}
}
\description{
For a given point finds the nearest point(s) to it within a given ball radius.
}
